refactor: extracted function into inline, move unit test to distributions brick tests
